+++ pkg/deployment/reconcile/action_wait_for_member_up.go
@@ -19,6 +19,7 @@
 //
 // Author Ewout Prangsma
 // Author Tomasz Mielech
+// Author Adam Janikowski
 //
 
 package reconcile
@@ -149,6 +150,8 @@ func (a *actionWaitForMemberUp) checkProgressAgent(ctx context.Context) (bool, b
 
 	shortCtx, c := context.WithTimeout(ctx, 3*time.Second)
 	defer c()
+
+	shortCtx = agency.WithAllowDifferentLeaderEndpoints(shortCtx)
 
 	if err := agency.AreAgentsHealthy(shortCtx, clients); err != nil {
 		log.Debug().Err(err).Msg("Not all agents are ready")